Eclipse errors cleanup
-Added ignore to non-supportable antrun task in maven lifecycle
-fixed what should be static references
-removed exception declarations from methods that don't throw that exception

+++ src/main/java/oscar/util/Pager.java
@@ -88,7 +88,7 @@ public class Pager {
 
             if (offset < (length - size)) {
                 header += ("&nbsp;<a href=\"" + url + pref + "pager.offset=" +
-                ((int) offset + (int) size) + "\">" + prop.getString("pager.next.desc") + "</a>\n");
+                (offset + size) + "\">" + prop.getString("pager.next.desc") + "</a>\n");
             }
 
             header += "</font>";
